backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg[binary]==3.1.18
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.1.4
django-cors-headers==4.3.1
google-auth==2.29.0
